pyproject.toml
[project]
name = "claw-easa"
version = "0.2.0"
description = "Query EASA Easy Access Rules — SQLite/FAISS skill for OpenClaw, hybrid retrieval, strict ref-only mode"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lxml>=6.0.2",
    "requests>=2.32.0",
    "beautifulsoup4>=4.14.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.25.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0"]
yaml = ["pyyaml>=6.0"]

[project.scripts]
claw-easa = "claw_easa.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claw_easa"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
